# openapi.yaml
openapi: 3.1.0
info:
  title: Checkout & Rewards Service
  version: 0.1.0
  description: |
    Backend for a small ecommerce store: products, carts, retry-safe checkout,
    orders, milestone coupons, and admin reporting.

    - Money is always integer cents (`long`).
    - Errors follow `{ code, message, details? }` with stable snake_case codes.
    - Checkout is retry-safe via the `Idempotency-Key` header.
    - Endpoints under `/admin/*` are the administrative surface. AuthN/AuthZ is
      not implemented in this take-home (see `DECISIONS.md`).
servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Products
  - name: Carts
  - name: Checkout
  - name: Orders
  - name: Admin

paths:
  /products:
    get:
      tags: [Products]
      summary: List catalog
      responses:
        "200":
          description: All products with current price and inventory.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
              examples:
                seeded:
                  value:
                    - { id: SKU-COFFEE,  name: "Single-origin coffee 250g", unitPriceCents: 1499,  inventory: 100 }
                    - { id: SKU-GRINDER, name: "Manual burr grinder",       unitPriceCents: 12900, inventory: 3   }

  /carts:
    post:
      tags: [Carts]
      summary: Create an empty cart
      responses:
        "201":
          description: New empty cart.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }

  /carts/{id}:
    get:
      tags: [Carts]
      summary: View cart with per-line current price, inventory, and totals
      parameters:
        - $ref: "#/components/parameters/CartId"
      responses:
        "200":
          description: Cart view.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404":
          $ref: "#/components/responses/CartNotFound"

  /carts/{id}/items:
    post:
      tags: [Carts]
      summary: Add a product to the cart (or add to an existing line's quantity)
      parameters:
        - $ref: "#/components/parameters/CartId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddItemRequest" }
            example: { productId: SKU-COFFEE, quantity: 2 }
      responses:
        "200":
          description: Updated cart.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "400":
          description: Validation error, unknown product, or invalid quantity.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                unknownProduct:
                  value: { code: unknown_product,  message: "product not found", details: { productId: SKU-NOPE } }
                invalidQuantity:
                  value: { code: invalid_quantity, message: "quantity must be >= 1" }
        "404":
          $ref: "#/components/responses/CartNotFound"
        "409":
          $ref: "#/components/responses/CartNotOpen"

  /carts/{id}/items/{productId}:
    put:
      tags: [Carts]
      summary: Replace the quantity of a line
      parameters:
        - $ref: "#/components/parameters/CartId"
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateItemRequest" }
            example: { quantity: 3 }
      responses:
        "200": { description: Updated cart., content: { application/json: { schema: { $ref: "#/components/schemas/Cart" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404":
          description: Cart or line not found.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                itemNotInCart:
                  value: { code: item_not_in_cart, message: "cart does not contain that product" }
        "409":
          $ref: "#/components/responses/CartNotOpen"
    delete:
      tags: [Carts]
      summary: Remove a line
      parameters:
        - $ref: "#/components/parameters/CartId"
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200": { description: Updated cart., content: { application/json: { schema: { $ref: "#/components/schemas/Cart" } } } }
        "404": { description: Cart or line not found. }
        "409": { $ref: "#/components/responses/CartNotOpen" }

  /carts/{id}/checkout:
    post:
      tags: [Checkout]
      summary: Validate the cart and place an order (optionally with a coupon)
      description: |
        Retry-safe. Callers should send an `Idempotency-Key` header for retries;
        a retry with the same key and the same body replays the stored response,
        and the same key with a different body returns `409 idempotency_key_reused`.
        As a belt-and-braces guarantee, a retry *without* a key on a cart that
        already produced an order returns that existing order rather than creating
        a second one.
      parameters:
        - $ref: "#/components/parameters/CartId"
        - name: Idempotency-Key
          in: header
          required: false
          description: Opaque, client-generated key (<= 128 chars). Recommended for all mutations.
          schema: { type: string, maxLength: 128 }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CheckoutRequest" }
            examples:
              withoutCoupon: { value: {} }
              withCoupon:    { value: { couponCode: "SAVE-ABCD1234" } }
      responses:
        "200":
          description: Order placed (or replay of an earlier successful order).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Empty cart or invalid coupon.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                emptyCart:      { value: { code: empty_cart,      message: "cart is empty" } }
                invalidCoupon:  { value: { code: invalid_coupon,  message: "coupon does not exist", details: { code: "NOPE" } } }
        "404":
          $ref: "#/components/responses/CartNotFound"
        "409":
          description: Any of several race/conflict outcomes; distinguish via the `code` field.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              examples:
                insufficientInventory:
                  value:
                    code: insufficient_inventory
                    message: "one or more products lack sufficient inventory"
                    details:
                      shortfalls:
                        - { productId: SKU-GRINDER, requested: 4, available: 3 }
                couponNotAvailable:
                  value:
                    code: coupon_not_available
                    message: "coupon is not available for redemption"
                    details: { code: "SAVE-ABCD1234", status: "REDEEMED" }
                idempotencyKeyReused:
                  value:
                    code: idempotency_key_reused
                    message: "Idempotency-Key was reused with a different request body"
                cartNotOpen:
                  value:
                    code: cart_not_open
                    message: "cart cannot be checked out"
                    details: { status: "CHECKED_OUT" }
                productMissing:
                  value:
                    code: product_missing
                    message: "cart references products that no longer exist"
                    details: { productIds: ["SKU-DELETED"] }

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Retrieve an order (immutable snapshot)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Order.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Order not found.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }

  /admin/coupons/generate:
    post:
      tags: [Admin]
      summary: Generate one coupon for the next reached-but-unrewarded milestone
      description: |
        For configuration `n=5, x=10`, every 5th successful order makes one 10%-off
        coupon available. This endpoint mints one coupon per call, starting at the
        smallest unrewarded milestone. Concurrent calls that race the same milestone
        are prevented by a UNIQUE constraint on `coupons.milestone` — only one wins.
      responses:
        "200":
          description: A new coupon was generated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerateCouponResponse" }
              example:
                { code: "SAVE-K7X2M9RQ", percent: 10, milestone: 5, status: "AVAILABLE", createdAt: "2026-09-20T04:00:00Z" }
        "409":
          description: No unrewarded milestone is eligible right now.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                code: no_milestone_available
                message: "no unrewarded milestone is eligible"
                details: { orderCount: 3, n: 5, eligibleMilestones: 0, couponsAlreadyGenerated: 0 }

  /admin/report:
    get:
      tags: [Admin]
      summary: Reconciled summary (read-only; repeated calls do not mutate state)
      responses:
        "200":
          description: Reconciled report.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Report" }

components:
  parameters:
    CartId:
      name: id
      in: path
      required: true
      description: Cart UUID.
      schema: { type: string, format: uuid }
    ProductId:
      name: productId
      in: path
      required: true
      description: Product SKU.
      schema: { type: string }

  responses:
    BadRequest:
      description: Validation error.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    CartNotFound:
      description: Cart not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { code: cart_not_found, message: "cart not found", details: { cartId: "..." } }
    CartNotOpen:
      description: Cart is not open for modification (already checked out).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { code: cart_not_open, message: "cart is not open for modification", details: { status: "CHECKED_OUT" } }

  schemas:
    Product:
      type: object
      required: [id, name, unitPriceCents, inventory]
      properties:
        id: { type: string }
        name: { type: string }
        unitPriceCents: { type: integer, format: int64, description: "Price in minor units (cents)." }
        inventory: { type: integer, minimum: 0 }

    AddItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: string }
        quantity: { type: integer, minimum: 1 }

    UpdateItemRequest:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer, minimum: 1 }

    CartLine:
      type: object
      properties:
        productId: { type: string }
        name: { type: string }
        unitPriceCents: { type: integer, format: int64 }
        quantity: { type: integer, minimum: 1 }
        lineTotalCents: { type: integer, format: int64 }
        availableInventory:
          type: integer
          nullable: true
          description: Current stock (may be null if the product was deleted after add).
        priceChangedSinceAdd:
          type: boolean
          description: Reserved for a future feature; currently always false.

    Cart:
      type: object
      required: [cartId, status, lines, subtotalCents]
      properties:
        cartId: { type: string, format: uuid }
        status:
          type: string
          enum: [OPEN, CHECKED_OUT, ABANDONED]
        lines:
          type: array
          items: { $ref: "#/components/schemas/CartLine" }
        subtotalCents: { type: integer, format: int64 }

    CheckoutRequest:
      type: object
      properties:
        couponCode:
          type: string
          nullable: true
          description: Optional coupon code, previously issued by `POST /admin/coupons/generate`.

    OrderLine:
      type: object
      required: [productId, productName, unitPriceCents, quantity, lineTotalCents]
      properties:
        productId: { type: string }
        productName: { type: string, description: "Snapshot at checkout time." }
        unitPriceCents: { type: integer, format: int64, description: "Snapshot at checkout time." }
        quantity: { type: integer, minimum: 1 }
        lineTotalCents: { type: integer, format: int64 }

    Order:
      type: object
      required: [orderId, cartId, placedAt, lines, grossCents, discountCents, netCents]
      properties:
        orderId: { type: string }
        cartId:  { type: string, format: uuid }
        placedAt: { type: string, format: date-time }
        lines:
          type: array
          items: { $ref: "#/components/schemas/OrderLine" }
        grossCents:    { type: integer, format: int64 }
        discountCents: { type: integer, format: int64 }
        netCents:      { type: integer, format: int64, description: "Always >= 0." }
        couponCode:    { type: string, nullable: true }
        couponPercent: { type: integer, nullable: true }

    GenerateCouponResponse:
      type: object
      required: [code, percent, milestone, status, createdAt]
      properties:
        code: { type: string }
        percent: { type: integer, minimum: 1, maximum: 100 }
        milestone: { type: integer, format: int64 }
        status: { type: string, enum: [AVAILABLE, REDEEMED] }
        createdAt: { type: string, format: date-time }

    CouponReportRow:
      type: object
      properties:
        code: { type: string }
        percent: { type: integer }
        milestone: { type: integer, format: int64 }
        status: { type: string, enum: [AVAILABLE, REDEEMED] }
        createdAt: { type: string, format: date-time }
        redeemedAt: { type: string, format: date-time, nullable: true }
        redeemedOnOrderId: { type: string, nullable: true }

    ProductSalesRow:
      type: object
      properties:
        productId: { type: string }
        name: { type: string }
        unitsSold: { type: integer, format: int64 }
        grossCents: { type: integer, format: int64 }

    Report:
      type: object
      required:
        [totalOrders, grossRevenueCents, totalDiscountsCents, netRevenueCents,
         salesByProduct, couponsGenerated, couponsAvailable, couponsRedeemed, coupons]
      properties:
        totalOrders:         { type: integer, format: int64 }
        grossRevenueCents:   { type: integer, format: int64 }
        totalDiscountsCents: { type: integer, format: int64 }
        netRevenueCents:     { type: integer, format: int64 }
        salesByProduct:
          type: array
          items: { $ref: "#/components/schemas/ProductSalesRow" }
        couponsGenerated: { type: integer, format: int64 }
        couponsAvailable: { type: integer, format: int64 }
        couponsRedeemed:  { type: integer, format: int64 }
        coupons:
          type: array
          items: { $ref: "#/components/schemas/CouponReportRow" }

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable snake_case slug clients can switch on.
          enum:
            - validation_error
            - bad_request
            - unknown_product
            - invalid_quantity
            - invalid_coupon
            - empty_cart
            - cart_not_found
            - item_not_in_cart
            - order_not_found
            - cart_not_open
            - insufficient_inventory
            - coupon_not_available
            - product_missing
            - idempotency_key_reused
            - no_milestone_available
        message: { type: string, description: "Human-readable, safe to expose." }
        details:
          type: object
          description: Structured extras (e.g. shortfalls, ids, current status).
          additionalProperties: true